Created separate index file with exports
